refactor: solve the problem with SharedModule & simplify import paths
