+++ src/app/core/index.ts
@@ -1,3 +1,2 @@
 export * from './enums';
 export * from './interceptors';
-export * from './guards';
