+++ src/app/features/content-controls/components/filter-bar/filter-bar.component.ts
