+++ src/app/features/content-controls/index.ts
@@ -1,3 +1,3 @@
-export * from './content-control.module';
+export * from './content-controls.module';
 export * from './components';
 export * from './enums';
