+++ src/app/features/features.module.ts
@@ -1,15 +1,17 @@
 import { NgModule } from "@angular/core";
-import { ContentControlModule } from "./content-control";
+import { ContentControlsModule } from "./content-controls";
 import { CourseModule } from "./course";
 import { FileModule } from "./file";
 import { TaskModule } from "./task";
+import { UserModule } from "./user";
 
 @NgModule({
   exports: [
     CourseModule,
     TaskModule,
     FileModule,
-    ContentControlModule
+    ContentControlsModule,
+    UserModule,
   ]
 })
 export class FeaturesModule {}
