+++ src/app/features/task/pipes/index.ts
@@ -1,0 +1,1 @@
+export * from './sort.pipe';
