+++ src/app/features/task/task.module.ts
@@ -1,14 +1,21 @@
 import { NgModule } from "@angular/core";
 import { ReactiveFormsModule } from "@angular/forms";
-import { SharedModule } from "../../shared";
+import { SharedModule } from "@shared";
 import { TaskItemComponent } from "./components";
+import { SortPipe } from "./pipes";
 
 @NgModule({
-  declarations: [TaskItemComponent],
+  declarations: [
+    TaskItemComponent,
+    SortPipe
+  ],
   imports: [
     SharedModule,
     ReactiveFormsModule
   ],
-  exports: [TaskItemComponent]
+  exports: [
+    TaskItemComponent,
+    SortPipe
+  ]
 })
 export class TaskModule {}
