+++ src/app/pages/auth/auth.module.ts
@@ -1,9 +1,6 @@
 import { AuthRoutingModule } from './auth-routing.module';
-import { AuthGuard, LoggedInGuard } from './guards';
-import { FlexLayoutModule } from '@angular/flex-layout';
 import { RouterModule } from '@angular/router';
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { LoginComponent, SignupComponent } from './components';
 import { ReactiveFormsModule } from '@angular/forms';
 import { SharedModule } from '@shared';
