--- src/app/pages/student/guards/index.ts
@@ -1,1 +1,0 @@
-export * from './course.guard';
