--- src/app/pages/user/guards/index.ts
@@ -1,1 +1,0 @@
-export * from './user.guard';
