+++ src/app/pages/user/user.module.ts
@@ -1,9 +1,15 @@
 import { NgModule } from '@angular/core';
+import { SharedModule } from '@shared';
+import { UserRoutingModule } from './user-routing.module';
 import { UserPageComponent } from './user.component';
 
 @NgModule({
   declarations: [
     UserPageComponent
+  ],
+  imports: [
+    SharedModule,
+    UserRoutingModule
   ]
 })
 export class UserPageModule {}
