+++ src/app/shared/guards/auth.guard.ts
