+++ src/app/shared/guards/course.guard.ts
