+++ src/app/shared/guards/index.ts
@@ -1,2 +1,4 @@
 export * from './auth.guard';
+export * from './course.guard';
 export * from './logged-in.guard';
+export * from './user.guard';
