+++ src/app/shared/guards/user.guard.ts
@@ -1,4 +1,4 @@
-import { AuthService } from '../../auth/services';
+import { AuthService } from '../services';
 import { Injectable } from '@angular/core';
 import { CanLoad, Route, UrlSegment, UrlTree } from '@angular/router';
 import { Observable } from 'rxjs';
@@ -7,7 +7,6 @@ import { Observable } from 'rxjs';
   providedIn: 'root'
 })
 export class UserGuard implements CanLoad {
-
   constructor(private authService: AuthService) {}
   canLoad(route: Route, segments: UrlSegment[]): boolean | UrlTree | Observable<boolean | UrlTree> | Promise<boolean | UrlTree> {
     return this.authService.isLoggedIn() ? true : false;
