+++ src/app/shared/pipes/pipes.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from "@angular/core";
+import { TruncatePipe } from "./truncate.pipe";
+
+@NgModule({
+  declarations: [TruncatePipe],
+  exports: [TruncatePipe]
+})
+export class PipesModule {}
