+++ src/app/shared/shared.module.ts
@@ -2,19 +2,14 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from "@angular/core";
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { MaterialModule } from './modules';
-import { SortPipe, TruncatePipe } from './pipes';
+import { PipesModule } from './pipes';
 
 @NgModule({
-  declarations: [
-    SortPipe,
-    TruncatePipe
-  ],
   exports: [
     CommonModule,
     MaterialModule,
     FlexLayoutModule,
-    SortPipe,
-    TruncatePipe
+    PipesModule,
   ]
 })
 export class SharedModule {}
